requirements.txt
# Core ML Libraries
scikit-learn==1.3.0
sentence-transformers==2.2.2
xgboost==1.7.6
ortools==9.7.2996

# Data Processing
pandas==2.0.3
numpy==1.24.3

# API and Web Framework
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0

# Database
sqlalchemy==2.0.21
sqlite3

# Visualization and Analysis
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
joblib==1.3.2
tqdm==4.65.0

# Data Generation (for testing)
faker==19.6.2

# Logging
loguru==0.7.0

# Configuration
pyyaml==6.0.1

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1
